+++ setup.cfg
@@ -76,6 +76,7 @@ tests =
     coverage
     black
     uvicorn  # move later
+    pystow
 pandas =
     pandas
 bioregistry =
