+++ tests/test_mapping_service.py
@@ -2,16 +2,12 @@
 
 """Tests for the identifier mapping service."""
 
-import subprocess
 import json
-import time
 import unittest
 from typing import Iterable, Set, Tuple
 from urllib.parse import quote
 from xml import etree
-from multiprocessing import Process
 
-import requests
 from fastapi.testclient import TestClient
 from rdflib import OWL, SKOS
 from rdflib.query import ResultRow
@@ -366,108 +362,3 @@ class TestFastAPIMappingApp(ConverterMixin):
     def test_post_service_query(self):
         """Test sparql generated by a service (that has values outside of where clause) with POST."""
         self.assert_post_sparql_results(self.client, SPARQL_FROM_SERVICE)
-
-
-
-# MAPPING_ENDPOINT = "https://bioregistry.io/sparql"
-MAPPING_ENDPOINT = "http://localhost:8000/sparql"
-BLAZEGRAPH_ENDPOINT = "http://localhost:9999/blazegraph/namespace/kb/sparql"
-
-FEDERATED_QUERY = f"""PREFIX owl: <http://www.w3.org/2002/07/owl#>
-SELECT DISTINCT ?o WHERE {{
-    SERVICE <{MAPPING_ENDPOINT}> {{
-        <http://purl.obolibrary.org/obo/CHEBI_24867> owl:sameAs ?o
-    }}
-}}"""
-
-# FEDERATED_QUERY = f"""PREFIX owl: <http://www.w3.org/2002/07/owl#>
-# SELECT DISTINCT ?o WHERE {{
-#     SERVICE <{MAPPING_ENDPOINT}> {{
-#         ?s ?p ?o
-#     }}
-# }}"""
-
-class TestFederatedSparql(unittest.TestCase):
-    """Test the identifier mapping service."""
-
-    def setUp(self):
-        import os
-        import uvicorn
-        # Start the mapping service endpoint in the background
-        converter = Converter.from_priority_prefix_map(PREFIX_MAP)
-        app = get_fastapi_mapping_app(converter)
-
-        # Download blazegraph.jar if not present
-        if not os.path.exists("blazegraph.jar"):
-            subprocess.run("wget -N https://github.com/blazegraph/database/releases/download/BLAZEGRAPH_2_1_6_RC/blazegraph.jar", shell=True)
-
-        # NOTE: Try using rdflib-endpoint instead of the curies mapping service
-        # import rdflib
-        # from rdflib_endpoint import SparqlEndpoint
-        # g = rdflib.Graph()
-        # g.add((rdflib.URIRef("http://s"), rdflib.URIRef("http://p"), rdflib.URIRef("http://o")))
-        # app = SparqlEndpoint(graph=g)
-
-        # Start the curies mapping service SPARQL endpoint
-        self.proc = Process(
-            target=uvicorn.run,
-            args=(
-                app,
-            ),
-            # kwargs={"host": "localhost", "port": 8000, "log_level": "info"},
-            daemon=True,
-        )
-        self.proc.start()
-
-        # Start Blazegraph SPARQL endpoint
-        self.proc_blazegraph = Process(
-            target=subprocess.run,
-            args=(
-                ["java -jar blazegraph.jar"]
-            ),
-            kwargs={
-                "shell": True,
-            },
-            daemon=True,
-        )
-        self.proc_blazegraph.start()
-        time.sleep(1)
-
-
-    def tearDown(self):
-        # TODO: blazegraph not actually killed, it will continue to run afterward
-        # Not sure how to make sure a java subprocess running as daemon is killed
-        self.proc_blazegraph.kill()
-        self.proc.kill()
-
-
-    def test_federated_local_blazegraph(self):
-        """Test sending a federated query to a local mapping service from a local Blazegraph."""
-
-        try:
-            resp = requests.get(
-                BLAZEGRAPH_ENDPOINT,
-                params={"query": FEDERATED_QUERY},
-                headers={"accept": "application/json"},
-            )
-            res = resp.json()
-            self.assertGreater(
-                len(res["results"]["bindings"]),
-                0,
-                msg=f"Federated query to {BLAZEGRAPH_ENDPOINT} gives no results",
-            )
-            return res["results"]["bindings"]
-        except Exception:
-            self.assertTrue(
-                False, msg=f"Error running the federated query to {BLAZEGRAPH_ENDPOINT}: {resp.text}"
-            )
-        return None
-
-        # NOTE: test using oxigraph as backend
-        # g = Graph(store="Oxigraph")
-        # app = SparqlEndpoint(graph=g)
-        # endpoint = TestClient(app)
-        # response = endpoint.get("/?query=" + urllib.parse.quote(FEDERATED_QUERY), headers={"accept": "application/json"})
-        # print(response.json()["results"]["bindings"])
-        # assert response.status_code == 200
-        # assert len(response.json()["results"]["bindings"]) > 0
